start settings migration
