+++ .vscode/settings.json
@@ -8,13 +8,8 @@
 
   "python.testing.unittestEnabled": false,
   "python.testing.nosetestsEnabled": false,
-  "python.discoverTest": true,
+  "python.testing.cwd": "./mealie/tests",
   "python.testing.pytestEnabled": true,
-  "cSpell.enableFiletypes": [
-    "!javascript",
-    "!python"
-  ],
-  "python.testing.pytestArgs": [
-    "mealie"
-  ]
+  "cSpell.enableFiletypes": ["!javascript", "!python"],
+  "python.testing.pytestArgs": ["mealie/test/"]
 }
