+++ mealie/test/conftest.py
@@ -1,0 +1,2 @@
+from pytest import fixture
+
