+++ mealie/test/pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+python_files = test_*
+python_classes = *Tests
+python_functions = test_*
